pyproject.toml
[tool.poetry]
name = "djangostripe"
version = "0.1.0"
description = ""
authors = ["Almaz Durusaliev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.1"
stripe = "^8.4.0"
django-environ = "^0.11.2"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"
pytest = "^8.0.2"
pytest-django = "^4.8.0"


[tool.poetry.group.linters.dependencies]
pre-commit = "^3.6.2"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
